pyproject.toml
[tool.poetry]
name = "mvxmt-backend"
version = "0.1.0"
description = ""
authors = ["ninjawarrior1337 <me@treelar.xyz>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pyjwt = "^2.9.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
ollama = "^0.4.2"
numpy = "^2.1.3"
more-itertools = "^10.5.0"
python-ulid = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
